#821 [REFACTOR] Options를 OptionModal에 응집
