+++ src/shared/component/index.js
@@ -26,7 +26,7 @@ export { default as FetchLoadingOverlay } from './loading/FetchLoadingOverlay/Fe
 // modal
 export { default as ConfirmModal } from './modal/ConfirmModal/ConfirmModal';
 export { default as DeleteModal } from './modal/DeleteModal';
-export { default as OptionModal } from './modal/OptionModal/OptionModal';
+export { default as OptionModal } from './modal/OptionModal';
 
 // suspense
 export { default as ServerErrorFallback } from './suspense/ServerErrorFallback/ServerErrorFallback';
