+++ src/shared/component/modal/OptionModal.jsx
@@ -1,5 +1,4 @@
-import Options from '@/shared/component/modal/OptionModal/Options';
-
+import { Icon } from '@/shared/component';
 import { MODAL_OPTIONS } from '@/constants';
 
 import styles from '@/shared/component/modal/Modal.module.css';
@@ -46,6 +45,48 @@ export default function OptionModal({
           </div>
         </div>
       </div>
+    </div>
+  );
+}
+
+function Options({ options, functions, onOptionClick }) {
+  const handleClick = (iconId) => {
+    if (iconId && functions !== undefined && functions[iconId]) {
+      functions[iconId]();
+    }
+  };
+
+  return (
+    <div className={styles.options}>
+      {options.map((option) => (
+        <div
+          className={styles.optionBar}
+          key={option.iconId}
+          data-value={option.value}
+          onClick={(event) => {
+            event.stopPropagation();
+            handleClick(option.iconId);
+
+            if (onOptionClick !== undefined) {
+              onOptionClick(event);
+            }
+          }}
+        >
+          <div className={styles.iconBox}>
+            <Icon
+              id={option.iconId}
+              width={option.IconWidth}
+              height={option.IconHeight}
+            />
+          </div>
+          <div
+            className={styles.optionText}
+            style={{ color: option.color || 'inherit' }}
+          >
+            {option.text}
+          </div>
+        </div>
+      ))}
     </div>
   );
 }
